定时任务更新每日节点年化率

+++ scan-service/src/main/java/com/platon/browser/service/staking/TxInfo.java
@@ -3,12 +3,13 @@ package com.platon.browser.service.staking;
 import lombok.Data;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @Data
 public class TxInfo{
 	private String nodeName;
 	private BigDecimal amount;
-	private int stakingBlockNum;
+	private BigInteger stakingBlockNum;
 	private int type;
 	private String nodeId;
 }
